example overhaul: modern_data_stack_assets to recommended project layout + rename 3.5/n
