+++ docs/content/concepts/assets/software-defined-assets.mdx
@@ -623,10 +623,10 @@ For more examples of software-defined assets, check out these examples:
   - [Per-asset IO manager](https://github.com/dagster-io/dagster/blob/master/examples/hacker_news_assets/hacker_news_assets/core/assets/items.py)
   - [Partitioned assets](https://github.com/dagster-io/dagster/blob/master/examples/hacker_news_assets/hacker_news_assets/core/assets/items.py)
 
-- In the [Modern Data Stack example](https://github.com/dagster-io/dagster/tree/master/examples/modern_data_stack_assets):
+- In the [Modern Data Stack example](https://github.com/dagster-io/dagster/tree/master/examples/assets_modern_data_stack):
 
-  - [Defining assets](https://github.com/dagster-io/dagster/blob/master/examples/modern_data_stack_assets/modern_data_stack_assets/assets.py)
-  - [Loading assets from dbt](https://github.com/dagster-io/dagster/blob/master/examples/modern_data_stack_assets/modern_data_stack_assets/assets.py)
-  - [Loading assets from Airbyte](https://github.com/dagster-io/dagster/blob/master/examples/modern_data_stack_assets/modern_data_stack_assets/assets.py)
+  - [Defining assets](https://github.com/dagster-io/dagster/blob/master/examples/assets_modern_data_stack/assets_modern_data_stack/assets/forecasting.py)
+  - [Loading assets from dbt](https://github.com/dagster-io/dagster/blob/master/examples/assets_modern_data_stack/assets_modern_data_stack/dbt.py)
+  - [Loading assets from Airbyte](https://github.com/dagster-io/dagster/blob/master/examples/assets_modern_data_stack/assets_modern_data_stack/airbyte.py)
 
 - The [Bollinger example](https://github.com/dagster-io/dagster/tree/master/examples/bollinger)
