+++ examples/assets_modern_data_stack/assets_modern_data_stack/__init__.py
@@ -1,0 +1,1 @@
+from .repository import assets_modern_data_stack
