+++ examples/assets_modern_data_stack/assets_modern_data_stack/utils/constants.py
@@ -1,17 +1,12 @@
-import numpy as np
 from dagster_postgres.utils import get_conn_string
 
 from dagster._utils import file_relative_path
 
 # =========================================================================
-# To get this value, run `python -m modern_data_stack_assets.setup_airbyte`
+# To get this value, run `python -m assets_modern_data_stack.setup_airbyte`
 # and grab the connection id that it prints at the end
 AIRBYTE_CONNECTION_ID = "your_airbyte_connection_id"
 # =========================================================================
-
-
-def model_func(x, a, b):
-    return a * np.exp(b * (x / 10**18 - 1.6095))
 
 
 PG_SOURCE_CONFIG = {
@@ -30,8 +25,8 @@ PG_DESTINATION_CONFIG = {
 }
 
 AIRBYTE_CONFIG = {"host": "localhost", "port": "8000"}
-DBT_PROJECT_DIR = file_relative_path(__file__, "../mds_dbt")
-DBT_PROFILES_DIR = file_relative_path(__file__, "../mds_dbt/config")
+DBT_PROJECT_DIR = file_relative_path(__file__, "../../mds_dbt")
+DBT_PROFILES_DIR = file_relative_path(__file__, "../../mds_dbt/config")
 DBT_CONFIG = {"project_dir": DBT_PROJECT_DIR, "profiles_dir": DBT_PROFILES_DIR}
 POSTGRES_CONFIG = {
     "con_string": get_conn_string(
