+++ examples/assets_modern_data_stack/mds_dbt/config/.user.yml
@@ -1,0 +1,1 @@
+id: 12ae8845-1563-4601-9113-11b298de3c6e
