+++ examples/assets_modern_data_stack/mds_dbt/dbt_project.yml
