+++ examples/assets_modern_data_stack/mds_dbt/models/daily_order_summary.sql
