+++ examples/assets_modern_data_stack/mds_dbt/models/orders_cleaned.sql
