+++ examples/assets_modern_data_stack/mds_dbt/models/schema.yml
