+++ examples/assets_modern_data_stack/mds_dbt/models/sources.yml
