+++ examples/assets_modern_data_stack/mds_dbt/models/users_augmented.sql
