+++ examples/assets_modern_data_stack/requirements.txt
@@ -1,0 +1,3 @@
+-e assets_modern_data_stack
+# It's recommended to include environment-specific requirements in this file.
+# For example, pinning a package's version for deployment.
