+++ examples/assets_modern_data_stack/setup.cfg
@@ -1,0 +1,2 @@
+[metadata]
+name = assets_modern_data_stack
