+++ examples/assets_modern_data_stack/tox.ini
@@ -2,8 +2,9 @@
 envlist = py{38, 37},pylint,mypy
 
 [testenv]
+usedevelop = true
 setenv =
-  VIRTUALENV_PIP=21.3.1
+  VIRTUALENV_PIP=22.1.2
 passenv = CI_* COVERALLS_REPO_TOKEN BUILDKITE
 deps =
   -e ../../python_modules/dagster[mypy,test]
@@ -25,4 +26,4 @@ commands =
 
 [testenv:pylint]
 commands =
-  pylint -j0 --rcfile=../../pyproject.toml {posargs} modern_data_stack_assets modern_data_stack_assets_tests
+  pylint -j0 --rcfile=../../pyproject.toml {posargs} assets_modern_data_stack assets_modern_data_stack_tests
