+++ examples/assets_modern_data_stack/workspace.yaml
@@ -1,0 +1,2 @@
+load_from:
+  - python_package: assets_modern_data_stack
