--- examples/modern_data_stack_assets/modern_data_stack_assets/__init__.py
@@ -1,1 +1,0 @@
-from .assets import mds_repo
