--- examples/modern_data_stack_assets/modern_data_stack_assets_tests/test_repo_loads.py
@@ -1,3 +1,0 @@
-def test_repo_loads():
-    # placeholder for future tests
-    return True
